Change how truck decide to disable or not signals
The `LISTENERS_OFF_BY_DEFAULT` variable is now a lazy accessor of the
`settings.TESTING` var. It's smart enough to detect when it should
disable the signals or not given the `TESTING` variable value;

If it is empty, does not exist or contain "acceptance", the flag is
False, the signals won't be disabled, otherwise the flag is set to true
and the signals will be disabled.